[FLINK-30659][hive] Drop flink-sql-parser-hive module
