--- flink-table/flink-sql-parser-hive/src/main/java/org/apache/flink/sql/parser/hive/ddl/SqlDescribeHiveTable.java
@@ -1,53 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.sql.parser.hive.ddl;
-
-import org.apache.flink.sql.parser.dql.SqlRichDescribeTable;
-
-import org.apache.calcite.sql.SqlIdentifier;
-import org.apache.calcite.sql.SqlWriter;
-import org.apache.calcite.sql.parser.SqlParserPos;
-
-/** DESCRIBE DDL to describe a Hive table. */
-public class SqlDescribeHiveTable extends SqlRichDescribeTable {
-
-    private final boolean extended;
-    private final boolean formatted;
-
-    public SqlDescribeHiveTable(
-            SqlParserPos pos,
-            SqlIdentifier tableNameIdentifier,
-            boolean extended,
-            boolean formatted) {
-        super(pos, tableNameIdentifier, extended || formatted);
-        this.extended = extended;
-        this.formatted = formatted;
-    }
-
-    @Override
-    public void unparse(SqlWriter writer, int leftPrec, int rightPrec) {
-        writer.keyword("DESCRIBE");
-        if (extended) {
-            writer.keyword("EXTENDED");
-        } else if (formatted) {
-            writer.keyword("FORMATTED");
-        }
-        tableNameIdentifier.unparse(writer, leftPrec, rightPrec);
-    }
-}
