+++ flink-table/flink-sql-parser/src/main/java/org/apache/flink/sql/parser/validate/FlinkSqlConformance.java
@@ -25,13 +25,7 @@ import org.apache.calcite.sql.validate.SqlConformanceEnum;
 /** Sql conformance used for flink to set specific sql dialect parser. * */
 public enum FlinkSqlConformance implements SqlConformance {
     /** Calcite's default SQL behavior. */
-    DEFAULT,
-
-    /**
-     * Conformance value that instructs Calcite to use SQL semantics consistent with the Apache
-     * HIVE, but ignoring its more inconvenient or controversial dicta.
-     */
-    HIVE;
+    DEFAULT;
 
     @Override
     public boolean isLiberal() {
@@ -60,24 +54,12 @@ public enum FlinkSqlConformance implements SqlConformance {
 
     @Override
     public boolean isSortByOrdinal() {
-        switch (this) {
-            case DEFAULT:
-            case HIVE:
-                return true;
-            default:
-                return false;
-        }
+        return false;
     }
 
     @Override
     public boolean isSortByAlias() {
-        switch (this) {
-            case DEFAULT:
-            case HIVE:
-                return true;
-            default:
-                return false;
-        }
+        return false;
     }
 
     @Override
@@ -137,11 +119,7 @@ public enum FlinkSqlConformance implements SqlConformance {
 
     @Override
     public boolean allowExplicitRowValueConstructor() {
-        switch (this) {
-            case DEFAULT:
-                return true;
-        }
-        return false;
+        return true;
     }
 
     @Override
