+++ flink-table/flink-table-planner/src/main/java/org/apache/flink/table/planner/delegation/FlinkSqlParserFactories.java
@@ -18,9 +18,9 @@
 
 package org.apache.flink.table.planner.delegation;
 
-import org.apache.flink.sql.parser.hive.impl.FlinkHiveSqlParserImpl;
 import org.apache.flink.sql.parser.impl.FlinkSqlParserImpl;
 import org.apache.flink.sql.parser.validate.FlinkSqlConformance;
+import org.apache.flink.table.api.TableException;
 
 import org.apache.calcite.sql.parser.SqlParserImplFactory;
 import org.apache.calcite.sql.validate.SqlConformance;
@@ -31,10 +31,10 @@ public class FlinkSqlParserFactories {
     private FlinkSqlParserFactories() {}
 
     public static SqlParserImplFactory create(SqlConformance conformance) {
-        if (conformance == FlinkSqlConformance.HIVE) {
-            return FlinkHiveSqlParserImpl.FACTORY;
-        } else {
+        if (conformance == FlinkSqlConformance.DEFAULT) {
             return FlinkSqlParserImpl.FACTORY;
+        } else {
+            throw new TableException("Unsupported SqlConformance: " + conformance);
         }
     }
 }
