+++ flink-table/pom.xml
@@ -47,7 +47,6 @@ under the License.
 		<module>flink-sql-gateway</module>
 		<module>flink-sql-client</module>
 		<module>flink-sql-parser</module>
-		<module>flink-sql-parser-hive</module>
 		<module>flink-table-code-splitter</module>
 		<module>flink-table-test-utils</module>
 	</modules>
